Renamed something.js to model-updater.js and changed it a bit.
